Moved CollidableObject from physics/2d to physics since its common to both 2d and 3d
