+++ CMakeLists.txt
@@ -25,7 +25,7 @@ if(PROJECT_IS_TOP_LEVEL)
             include/physics/2d/Hitbox.h
             src/physics/2d/Collision.cpp
             include/physics/2d/Collision.h
-            include/physics/2d/CollidableObjectType.h
+            include/physics/CollidableObjectType.h
             include/physics/2d/CollisionAxis.h
         include/entity/player/sprites/PlayerSpriteState.h
         src/entity/player/sprites/PlayerSpriteHandler.cpp
@@ -125,7 +125,7 @@ add_executable(2d-platformer-sanitized main.cpp
         include/physics/2d/Hitbox.h
         src/physics/2d/Collision.cpp
         include/physics/2d/Collision.h
-        include/physics/2d/CollidableObjectType.h
+        include/physics/CollidableObjectType.h
         include/physics/2d/CollisionAxis.h
         include/entity/player/sprites/PlayerSpriteState.h
         src/entity/player/sprites/PlayerSpriteHandler.cpp
@@ -212,7 +212,7 @@ add_executable(2d-platformer-tsan main.cpp
         include/physics/2d/Hitbox.h
         src/physics/2d/Collision.cpp
         include/physics/2d/Collision.h
-        include/physics/2d/CollidableObjectType.h
+        include/physics/CollidableObjectType.h
         include/physics/2d/CollisionAxis.h
         include/entity/player/sprites/PlayerSpriteState.h
         src/entity/player/sprites/PlayerSpriteHandler.cpp
@@ -301,7 +301,7 @@ add_executable(performance_test performance_test.cpp
         include/physics/2d/Hitbox.h
         src/physics/2d/Collision.cpp
         include/physics/2d/Collision.h
-        include/physics/2d/CollidableObjectType.h
+        include/physics/CollidableObjectType.h
         include/physics/2d/CollisionAxis.h
         include/entity/player/sprites/PlayerSpriteState.h
         src/entity/player/sprites/PlayerSpriteHandler.cpp
@@ -384,7 +384,7 @@ add_library(game_lib STATIC
         include/physics/2d/Hitbox.h
         src/physics/2d/Collision.cpp
         include/physics/2d/Collision.h
-        include/physics/2d/CollidableObjectType.h
+        include/physics/CollidableObjectType.h
         include/physics/2d/CollisionAxis.h
         include/entity/player/sprites/PlayerSpriteState.h
         src/entity/player/sprites/PlayerSpriteHandler.cpp
